feat: add ui to workout and exercise details
