+++ composeApp/src/commonMain/composeResources/values/strings.xml
@@ -7,7 +7,6 @@
     <string name="select_measurement">Please select a measurement for your </string>
     <string name="height">Height</string>
     <string name="weight">Weight</string>
-
     <string name="start_now">Start Now</string>
     <string name="next">Next</string>
     <string name="internet_is_not_available">Internet is not available</string>
@@ -165,7 +164,14 @@
     <string name="workout_report">Workout Report</string>
     <string name="ic_arrow_down">Arrow Down Icon</string>
     <string name="activity">Activity</string>
-
     <string name="start_workout">Start Workout</string>
     <string name="close">Close</string>
+    <string name="second">Second</string>
+    <string name="detail_icon">detail icon</string>
+    <string name="focus_quadriceps">Quadriceps</string>
+    <string name="focus_abs">Abs</string>
+    <string name="focus_calves">Calves</string>
+    <string name="focus_lower_back">Lower Back</string>
+    <string name="focus_core">Core</string>
+    <string name="focus_shoulders">Shoulders</string>
 </resources>
