+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/viewmodel/workoutDetails/WorkoutDetailsEffect.kt
@@ -1,6 +1,5 @@
 package com.cairosquad.evolvefit.viewmodel.workoutDetails
 
-import com.cairosquad.evolvefit.viewmodel.register.RegisterEffect
 
 sealed class WorkoutDetailsEffect {
     object NavigateToPlayWorkout: WorkoutDetailsEffect()
