+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/viewmodel/workoutDetails/WorkoutDetailsInteractionListener.kt
@@ -1,10 +1,11 @@
 package com.cairosquad.evolvefit.viewmodel.workoutDetails
 
+
 interface WorkoutDetailsInteractionListener {
     fun onBackClick()
     fun onShareClick()
-    fun onAddToFavoriteClick()
-    fun onExerciseClick(exerciseName: String)
+    fun onAddToFavoriteClick(workoutId: String)
+    fun onExerciseClick(exercise: WorkoutDetailsScreenState.ExerciseUiState)
     fun onExerciseBottomSheetDismiss()
-    fun onStartWorkoutClick()
+    fun onStartWorkoutClick(workoutId: String)
 }
